Fix vc compilation error+warnings
git-svn-id: http://svn.cern.ch/guest/lhcb/Lbcom/trunk@89959 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 2224777f7784146f43ef3d60942c631e41e00955

+++ Lbcom/ST/STTools/src/STCMSNoiseCalculationTool.cpp
@@ -440,7 +440,7 @@ void ST::STCMSNoiseCalculationTool::sumPedestals(const LHCb::STTELL1Datas* data 
 	
         int iPort;
         for ( iPort = 0; iPort < 4; iPort++ ) {
-          int header = m_portHeader[iPort];
+          unsigned long header = m_portHeader[iPort];
           unsigned int iStrip = 0;
           for ( ; iStrip < LHCbConstants::nStripsInPort ; ++iStrip) {
             int strip = iStrip + iPort*LHCbConstants::nStripsInPort + beetle*LHCbConstants::nStripsInBeetle;
